requirements.txt
fair-esm
datasets
evaluate 
seqeval
ipython
ipywidgets
loguru
pandas
biopython
biotite
gemmi
scikit-learn
rich
transformers==4.56.2
accelerate
timm
plotly
tensorboard_plugin_profile
wandb
cupy-cuda11x
einops
tensorboard
torch-tb-profiler
lightning
pyrosetta-distributed 
pyrosetta-installer 
deepspeed
SentencePiece
e3nn
freesasa
torchsummary
func_timeout
torch_geometric
matplotlib
seaborn
cmake
tabulate
streamlit
biometall
allmetal3d
line_profiler
git+https://github.com/TurtleTools/portein
git+https://github.com/openmm/pdbfixer.git
git+https://github.com/openmm/openmmforcefields
--find-links https://data.pyg.org/whl/torch-2.7.0+cu118.html
torch-scatter
torch-sparse
torch-cluster
pyg_lib
ipywidgets
ipykernel